frontend/functions/requirements.txt
firebase_functions~=0.1.0
librosa==0.10.2.post1
numpy==2.0.0
soundfile==0.12.1
ffmpeg-python==0.2.0
google-cloud-storage==2.17.0
firebase-admin==6.5.0
